requirements.txt
python-telegram-bot==21.6
requests==2.32.3
apscheduler==3.10.4
feedparser==6.0.11
beautifulsoup4==4.12.3
google-generativeai==0.7.2
python-dotenv==1.0.1
lxml==5.2.2
